requirements.txt
streamlit
pandas
plotly
websockets
websocket-client